refactor: 통합 테스트 분리
